+++ components/draft/filehandler/functions/get/collectTieup.ts
