+++ components/draft/filehandler/functions/get/collectWarp.ts
@@ -1,4 +1,4 @@
-import { getHighest } from './utils'
+import { getHighest } from '../utils'
 
 export function collectWarp(warp: grid) {
