+++ components/draft/filehandler/functions/get/createWeaveObject.ts
